+++ lib/fishstick_dart.dart
@@ -13,6 +13,7 @@ export "commands/blacklist/blacklist.dart";
 export "commands/fortnite/account/login.dart";
 export "commands/fortnite/account/logout.dart";
 export "commands/fortnite/account/account.dart";
+export "commands/fortnite/account/launch.dart";
 export "commands/fortnite/vbucks/vbucks.dart";
 export "commands/fortnite/sac/sac.dart";
 export "commands/fortnite/mfa/mfa.dart";